Empty Functionality
Added functions to get data from the
api call and then iterate through it
recursively to check its arrays. if an
array is empty, its preprty or datatype is
added to an array. If it has data it is
added to another array. Based on these 2 arrays,
displays are shown or hidden.